+++ intent_QA/backup_20240530064051/Loki_classes.py
