+++ intent_QA/backup_20240530064051/Loki_reservation.py
