+++ intent_QA/intent_new/Loki_classes.py
